Worked on normalized points
Made it so when the user clicks anywhere on the image the coordinate and normalized coordinate are displayed on screen.

Currently the origin point of the image is set to the top left of the image.

Added a dropbox field so that the user can pick where the origin point is located. For now there will only be two options, Top-Left and Center.