Finished
Finished coding the project, removed the text displaying the points as the list was doing that anyway. Made it so the JSON  file downloaded is based on the drop-down menu. Also added a reset points whenever the image is change and the button is pressed. Also added cute styling to the site.